Fixed bandwidth calculations

+++ android/src/main/java/com/xh/networkspeed/NetworkSpeed.java
@@ -4,24 +4,24 @@ import android.net.TrafficStats;
 
 public class NetworkSpeed {
     private static final String TAG = NetworkSpeed.class.getSimpleName();
-    private long lastTotalRxBytes = 0;
-    private long lastTimeStamp = 0;
+    private double lastTotalRxBytes = 0;
+    private double lastTimeStamp = 0;
 
-    private long lastDownLoadBytes = 0;
-    private long lastUpLoadBytes = 0;
-    private long lastDownLoadBytesUid = 0;
-    private long lastUpLoadBytesUid = 0;
+    private double lastDownLoadBytes = 0;
+    private double lastUpLoadBytes = 0;
+    private double lastDownLoadBytesUid = 0;
+    private double lastUpLoadBytesUid = 0;
 
-    private long lastDownLoadTimeStamp = 0;
-    private long lastUpLoadTimeStamp = 0;
-    private long lastUpLoadTimeStampUid = 0;
-    private long lastDownLoadTimeStampUid = 0;
+    private double lastDownLoadTimeStamp = 0;
+    private double lastUpLoadTimeStamp = 0;
+    private double lastUpLoadTimeStampUid = 0;
+    private double lastDownLoadTimeStampUid = 0;
 
     public NetSpeedResult getNetSpeed(int uid) {
-        long downLoadSpeed = getTotalRxBytes();
-        long upLoadSpeed = getTotalTxBytes();
-        long downLoadSpeedUid = getTotalRxBytesByUid(uid);
-        long upLoadSpeedUid = getTotalTxBytesByUid(uid);
+        double downLoadSpeed = getTotalRxBytes();
+        double upLoadSpeed = getTotalTxBytes();
+        double downLoadSpeedUid = getTotalRxBytesByUid(uid);
+        double upLoadSpeedUid = getTotalTxBytesByUid(uid);
         NetSpeedResult result = new NetSpeedResult();
         result.setDownLoadSpeed(String.valueOf(downLoadSpeed));
         result.setDownLoadSpeedUid(String.valueOf(downLoadSpeedUid));
@@ -54,20 +54,20 @@ public class NetworkSpeed {
     }
 
     // 获取指定进程的发送流量
-    public long getTotalTxBytesByUid(int uid) {
-        long nowTimeStamp = System.currentTimeMillis();
-        long nowTotalTxBytes = TrafficStats.getUidTxBytes(uid) == TrafficStats.UNSUPPORTED ? 0 : (TrafficStats.getUidTxBytes(uid) / 1024);//转为KB
-        long speed = ((nowTotalTxBytes - lastUpLoadBytesUid) * 1000 / (nowTimeStamp - lastUpLoadTimeStampUid));//毫秒转换
+    public double getTotalTxBytesByUid(int uid) {
+        double nowTimeStamp = System.currentTimeMillis();
+        double nowTotalTxBytes = TrafficStats.getUidTxBytes(uid) == TrafficStats.UNSUPPORTED ? 0 : (TrafficStats.getUidTxBytes(uid) / 1024);//转为KB
+        double speed = ((nowTotalTxBytes - lastUpLoadBytesUid) * 1000 / (nowTimeStamp - lastUpLoadTimeStampUid));//毫秒转换
         lastUpLoadTimeStampUid = nowTimeStamp;
         lastUpLoadBytesUid = nowTotalTxBytes;
         return speed;
     }
 
     // 获取手机所有的发送流量
-    public long getTotalTxBytes() {
-        long nowTimeStamp = System.currentTimeMillis();
-        long nowTotalTxBytes = TrafficStats.getTotalTxBytes() == TrafficStats.UNSUPPORTED ? 0 : (TrafficStats.getTotalTxBytes() / 1024);//转为KB
-        long speed = ((nowTotalTxBytes - lastUpLoadBytes) * 1000 / (nowTimeStamp - lastUpLoadTimeStamp));//毫秒转换
+    public double getTotalTxBytes() {
+        double nowTimeStamp = System.currentTimeMillis();
+        double nowTotalTxBytes = TrafficStats.getTotalTxBytes() == TrafficStats.UNSUPPORTED ? 0 : (TrafficStats.getTotalTxBytes() / 1024);//转为KB
+        double speed = ((nowTotalTxBytes - lastUpLoadBytes) * 1000 / (nowTimeStamp - lastUpLoadTimeStamp));//毫秒转换
         lastUpLoadTimeStamp = nowTimeStamp;
         lastUpLoadBytes = nowTotalTxBytes;
         return speed;
